Logistic regression with confusion matrix
LR was explored as part of my thesis, to solve a problem. This confusion matrix was used to compare the performance of ML algorithms. Data for this experiment is confidential, but contains 12 different classes. I try to find the best accuracy of LR by tweaking the parameters and use the best parameters to print the confusion matrix.